refactor(pipelines/cdp): modularize test suite to mirror pipeline structure
- Decompose monolithic test_customer_data_platform.py into dedicated test modules:
  - test_models.py: tests for NamedTuple data models, payload parsing, and Beam schema compatibility
  - test_options.py: tests for pipeline and GCP options parsing and defaults
  - test_schemas.py: tests for packaged and custom BigQuery table schema loading
  - test_parsing.py: tests for ParseRecordDoFn and AssignEventTimestampDoFn
  - test_sessionization.py: tests for ProcessCustomerSessionDoFn aggregation and metrics
  - test_sinks.py: tests for Storage Write API timestamp formatting and BigQuery sink attachment
  - test_pipeline.py: tests for build_pipeline DAG assembly and in-memory execution
- Expand unit test coverage for sinks and timestamp assignment (41 tests passing, up from 21)
- Verify full compliance with Google Python style (yapf) and PyLint (10.00/10)

TAG=agy
CONV=59306c7c-80e2-4454-9ec5-e496e3b85156
